+++ sdk/storage/Azure.Storage.Blobs/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added support for service version 2020-08-04.
 - Added support for Blob Query Parquet input format.
 - Fixed bug where clients would sometimes throw a NullReferenceException when calling GenerateSas() with a BlobSasBuilder parameter.
+- Fixed bug where BlobBaseClient.Exists() would not function correctly on blobs encrypted with CPK.
 
 ## 12.9.0-beta.3 (2021-04-09)
 - This release contains bug fixes to improve quality.